pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolienas",email = "nicholienass@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "qrcode (>=8.2,<9.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "cairosvg (>=2.8.2,<3.0.0)",
    "cairocffi (>=1.7.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "reportlab (>=4.4.3,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "xlsxwriter (>=3.2.5,<4.0.0)",
    "amqp (>=5.3.1,<6.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
django = "^5.2.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
